+++ code/webapp/app/Http/Controllers/MentorController.php
@@ -7,6 +7,7 @@ use App\Models\Role;
 use App\Models\User;
 use App\Models\UserType;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class MentorController extends Controller
 {
@@ -34,6 +35,9 @@ class MentorController extends Controller
      */
     public function create()
     {
+
+        Gate::authorize('createDestroyTable');
+
         // send departments to the view
         $departments = Department::all();
         $roles = Role::all();
@@ -48,6 +52,8 @@ class MentorController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('createDestroyTable');
+
         $request->request->add(['user_type_id' => 3]);
         $request->request->add(['password' => bcrypt('password')]);
         User::create($request->all());
@@ -75,6 +81,7 @@ class MentorController extends Controller
      */
     public function edit($id)
     {
+        Gate::authorize('editUser', $id);
         $mentor = User::find($id);
         $departments = Department::all();
         $roles = Role::all();
@@ -90,6 +97,7 @@ class MentorController extends Controller
      */
     public function update(Request $request, $id)
     {
+        Gate::authorize('editUser', $id);
         $mentor = User::find($id);
         $mentor->update($request->all());
 
@@ -105,6 +113,8 @@ class MentorController extends Controller
      */
     public function destroy($id)
     {
+        Gate::authorize('createDestroyTable');
+        
         $mentor = User::find($id);
         $mentor->delete();
